We've fixed the bug and resubmitted. The artifact and version is "hw2-team08-0.0.13".
We re-organized the code of our retrieval part. We use the factory
module and added some syntax to our yaml file. So we don't need to
manually write code to initialize each query expander now. 

Besides, as we find the bug is caused by the third party resources we
used, we also modified the way each component access external resources.


One additional modification is that as the default simple bio passage
extractor produces better result than any of our implementations, we use
the default one as the passage extractor.